pyproject.toml
[tool.poetry]
name = "stoch3com"
version = "1.0.0"
description = ""
authors = ["Vitaliy Klychkov, Mykola Symon"]
license = "GPL-3.0 license"
readme = "README.md"
packages = [{include = "bot"}]

[tool.poetry.dependencies]
python = ">=3.8,<3.11"
python-dotenv = "^0.21.0"
python-telegram-bot = "^13.14"
ccxt = "^2.0.90"
pandas_ta = "*"
schedule = "^1.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
bot = "bot.main:main"
